# hoteli-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hoteli API Dokumentacioni",
        "description": "Dokumentacioni i API-së për Sistemin e Rezervimeve të Hotelit",
        "contact": {
            "email": "support@hoteli.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/receptionists": {
            "post": {
                "tags": [
                    "Receptionists"
                ],
                "summary": "Krijo një recepsionist të ri",
                "description": "Krijon një përdorues të ri me rolin 'receptionist'. Kërkon autentifikim.",
                "operationId": "storeReceptionist",
                "requestBody": {
                    "description": "Të dhënat e recepsionistit të ri",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Filan Fisteku"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "filan.fisteku@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "044123456"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recepsionisti u krijua me sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recepsionisti u krijua me sukses."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Të dhëna të pavlefshme",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pa autentifikim"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/receptionists/{receptionist}": {
            "put": {
                "tags": [
                    "Receptionists"
                ],
                "summary": "Përditëso një recepsionist ekzistues",
                "description": "Përditëson të dhënat e një recepsionisti. Kërkon autentifikim.",
                "operationId": "updateReceptionist",
                "parameters": [
                    {
                        "name": "receptionist",
                        "in": "path",
                        "description": "ID e recepsionistit për t'u përditësuar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Të dhënat e përditësuara të recepsionistit",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "is_active"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Filan Fisteku i Ri"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "filan.fisteku.ri@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "045987654"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword",
                                        "nullable": true
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recepsionisti u përditësua me sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recepsionisti u përditësua me sukses."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recepsionisti nuk u gjet"
                    },
                    "422": {
                        "description": "Të dhëna të pavlefshme"
                    },
                    "401": {
                        "description": "Pa autentifikim"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Receptionists"
                ],
                "summary": "Fshi një recepsionist",
                "description": "Fshin një recepsionist specifik. Kërkon autentifikim.",
                "operationId": "deleteReceptionist",
                "parameters": [
                    {
                        "name": "receptionist",
                        "in": "path",
                        "description": "ID e recepsionistit për t'u fshirë",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recepsionisti u fshi me sukses (No Content)"
                    },
                    "404": {
                        "description": "Recepsionisti nuk u gjet"
                    },
                    "401": {
                        "description": "Pa autentifikim"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}